Initial Swift 2.0 compatibility.
* Swift 2.0 compatibility.
* Extracted content views to separate classes.
* Minor cleanup.
