+++ PKHUD/PKHUDStatusVIew.swift
@@ -1,0 +1,64 @@
+//
+//  PKHUDStatusVIew.swift
+//  PKHUD
+//
+//  Created by Philip Kluz on 6/12/15.
+//  Copyright © 2015 NSExceptional. All rights reserved.
+//
+
+import Foundation
+
+/// PKHUDStatusView provides a square view, which you can use to display a picture, a title and a subtitle. This type of view replicates the Apple HUD one to one.
+public final class PKHUDStatusView: PKHUDImageView {
+    
+    public init(title: String?, subtitle: String?, image: UIImage?) {
+        super.init(image: image)
+        commonInit(title: title, subtitle: subtitle)
+    }
+    
+    public required init(coder aDecoder: NSCoder) {
+        super.init(coder: aDecoder)
+        commonInit(title: "", subtitle: "")
+    }
+    
+    private func commonInit(title title: String?, subtitle: String?) {
+        titleLabel.text = title
+        subtitleLabel.text = subtitle
+        
+        addSubview(titleLabel)
+        addSubview(subtitleLabel)
+    }
+    
+    public override func layoutSubviews() {
+        super.layoutSubviews()
+        
+        let viewWidth = bounds.size.width
+        let viewHeight = bounds.size.height
+        
+        let halfHeight = CGFloat(ceilf(CFloat(viewHeight / 2.0)))
+        let quarterHeight = CGFloat(ceilf(CFloat(viewHeight / 4.0)))
+        let threeQuarterHeight = CGFloat(ceilf(CFloat(viewHeight / 4.0 * 3.0)))
+        
+        titleLabel.frame = CGRect(origin: CGPointZero, size: CGSize(width: viewWidth, height: quarterHeight))
+        imageView.frame = CGRect(origin: CGPoint(x:0.0, y:quarterHeight), size: CGSize(width: viewWidth, height: halfHeight))
+        subtitleLabel.frame = CGRect(origin: CGPoint(x:0.0, y:threeQuarterHeight), size: CGSize(width: viewWidth, height: quarterHeight))
+    }
+    
+    public let titleLabel: UILabel = {
+        let label = UILabel()
+        label.textAlignment = .Center
+        label.font = UIFont.boldSystemFontOfSize(17.0)
+        label.textColor = UIColor.blackColor().colorWithAlphaComponent(0.85)
+        return label
+    }()
+    
+    public let subtitleLabel: UILabel = {
+        let label = UILabel()
+        label.textAlignment = .Center
+        label.font = UIFont.systemFontOfSize(14.0)
+        label.textColor = UIColor.blackColor().colorWithAlphaComponent(0.7)
+        label.adjustsFontSizeToFitWidth = true
+        label.numberOfLines = 2
+        return label
+    }()
+}
